Adding two sanity checks.
First one is for testing if the system RHS matrix is anti-symmetric.
This uses Petsc's 'MatisTranspose' routine to test if A=A^T and B=-B^T.

The second one tests if the vectorized state x=[u,v] represents a
hermitian matrix. For this to be true, u needs to represent a symmetric
matrix, v needs to represent an antisymmetric matrix. Vectorization is
column-wise.