Initialize the braid history in function InitGrids().
Do this before before braid_Drive is called (setting warm_restart to 1)
so that the first call to braid_Drive doesn't call the user's myInit
function. This is neccessary because the myInit function of the adjoint
solver does not contain the correct derivative. Derivative for adjoint
is set in adjoint->preprocess.