Fix groundstate gate.
The groundstate gate is NOT a linear transformation of the initial
conditions! Instead, it should drive any state towards the ground state,
being the first unit vector e_1 = (1 0 0 0 0 0 ...)^T.

Therefore, the groundstate gate computes the frobenius norm between the
end state x and the unit vector e_1, regardless of the initial
condition. To realize that, the gate routine 'compare' has been
overloaded for the groundstate gate.