Bugfix in parallel ReducedDens: MPI_Allreduce the values.
This is not a preferred solution, however it works:
The reduced density matrix is created as a sequential matrix (because
parallel layout would be different from full state). Values are set in a
loop that gets values from the full density matrix. The resulting
sequential vector is different on all processors, hence needs to be
summed up from all processors.