Init optim params s.t. 0 = control(0)=control(T).
For each oscillator, the control functions should start and end at 0.0.
The optimizer ensures that through the bounds, however the initial guess
did not yet satisfy this when initializing either randomly or with
constant amplitude. Now, the first and last two spline coefficients are
always set to zero so that the control functions start and end at zero
always.