Restructure evalF, evalGradF, objective, gate fidelity.
* Hilbert-Schmidt overlap returns a complex number Tr(state *
  target^dagger), with Im(HS-overlap) != 0 only if Schroedinger!

* evalJ in optimproblem also returns a complex number, with Im(J) != 0
  only if Schroedinger and J=Jtrace.

* Take the absolute values (or real parts) for the objective function
  and the fidelity only after all initial conditions are propagated.

* obj_weights are rescaled so that they always sum to 1.

* evalF:
Propagate *all* initial conditions forward to final time, while
summing up the *complex* values for the Hilbert-Schmidt overlap.
The absolute value is taken *after* the loop with a call to finalizeJ().

* evalGradF:
If Schroedinger, we need to backpropagate the adjoint variables only
after *all* the initial conditions have been propagated to final time T
(after finalizeJ()). Therefore the Schroedinger adjoint sweep
is performed in a separate loop after all primal states are computed.
The Schroedinger adjoint does *not* store the trajectories of the primal
state, but instead recomputes them backwards in time alongside the
adjoint variables.