Enforces control boundaries to start and end at zero.
Boundaries are enforced for the splines when setting the control
amplitudes from within the master equation. It overwrites the
optimization vector Vec x with zeros at the positions that correspond to
control function boundaries (i.e. the first and last two splines for
each carrier wave are set to zero.)