I partially documented the learning class I added documentation to setLearnParams, getLearnParams, and initLearnParams (the functions I need to  modify to make the transfer learning class have an arbitrary number of parameters per oscillator, which is what I am trying to implement.
I couldn't really follow a lot of the code (at least not as easily as I wanted to; I do better reading english than code), so I documented these files in addition to modifying them.

I am currently making changes to the TransferLearning class in UDEModel (to allow for a more general model).