Added more comment changes to code + got some stuff running in the python script I reformatted/rewrote a lot of the transferlearning class functions (in UDEmodel.py). I also made some more tweaks to the function signatures in learning.cpp (specify input our output for each argument). I plan to slowly make these chantges ot the rest of the code, as I get there.
The code now passes the first test (loss is zero when we use the same transfer function parameters). This is good; the issue before was that I only set up one set of offset/scale parameters... even though there are 2 carrier waves per oscillator. I added code that should detect the number of carrier waves per oscillator + use that info to construct the learn_params vector that we use to initialize the transfer model.