Compute trajectory output only if requested in the config.
Previously, trajectory output for the expected energy level and the
population were computed even though they were not written to files,
which increases compute time because this is done at each time step.
Now, the values are only computed if the user had requested this output
in the configuration file.

The configuration string is parsed in the constructor of the Output
class, which stores booleans to determine which output was requested by
the user. If booleans are true, the 0-th petsc rank opens the files,
writes the data per timestep, and closes the files.

Note that computing reduced output values per oscillator require an MPI
allreduce over all petsc ranks. Hence, all petsc ranks need to compute
the output whereas only the 0-th rank writes them to the open file.